+++ src/test/java/ru/yandex/practicum/filmorate/InMemoryUserManagerTest.java
@@ -4,7 +4,6 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.storage.implementation.InMemoryUserStorage;
 import ru.yandex.practicum.filmorate.model.User;
